% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argumentFormatter}
\alias{argumentFormatter}
\title{generate string that can be used as arguments}
\usage{
argumentFormatter(field, item, object, dbc)
}
\arguments{
\item{field}{which db field is used}

\item{item}{the item itself}

\item{object}{object it is associated with}

\item{dbc}{RangoConnection}
}
\value{
string that can be used as arguments
}
\description{
generate string that can be used as arguments
}
\author{
Willem Ligtenberg
}
