% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangoConnection.R
\name{beginTransaction}
\alias{beginTransaction}
\title{Begin an new transaction}
\usage{
beginTransaction(object)
}
\arguments{
\item{object}{the database connection to use}
}
\description{
Begin an new transaction
}
\author{
Willem Ligtenberg
}
