% Generated by roxygen2 (4.0.2): do not edit by hand
\name{commit}
\alias{commit}
\title{Commit the changes in the current transaction}
\usage{
commit(object)
}
\arguments{
\item{object}{the database connection to use}
}
\description{
Commit the changes in the current transaction
}
\author{
Willem Ligtenberg
}

