% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createObjectFromResult}
\alias{createObjectFromResult}
\title{Create an object from the result of the query}
\usage{
createObjectFromResult(object, dbc, result, create)
}
\arguments{
\item{object}{The object that was stored or retrieved}

\item{dbc}{database connection to use}

\item{result}{the result of the query}

\item{create}{Should we create the object if it did not exist?}
}
\value{
Object itself
}
\description{
Create an object from the result of the query
}
\author{
Willem Ligtenberg
}

