% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataType}
\alias{dataType}
\title{Detect the data type of the argument}
\usage{
dataType(columnDataRow)
}
\arguments{
\item{columnDataRow}{specific row of the description of a table}
}
\value{
data type as a string
}
\description{
Detect the data type of the argument
}
\author{
Willem Ligtenberg
}

