% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forceAssignMyNamespace}
\alias{forceAssignMyNamespace}
\title{Function to force the assignment of new items in the Rango namespace
Adapted from Rcpp}
\usage{
forceAssignMyNamespace(x, value)
}
\arguments{
\item{x}{name of the symbol/function inside a string}

\item{value}{new value of the symbol/function}
}
\description{
Function to force the assignment of new items in the Rango namespace
Adapted from Rcpp
}
\author{
Willem Ligtenberg
}
