% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formatter}
\alias{formatter}
\title{generate correct string representation of the object}
\usage{
formatter(object, dbc)
}
\arguments{
\item{object}{object to format}

\item{dbc}{RangoConnection}
}
\value{
string representation of the object
}
\description{
generate correct string representation of the object
}
\author{
Willem Ligtenberg
}

