% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{generateClass,RangoSQLiteConnection-method}
\alias{generateClass,RangoSQLiteConnection-method}
\title{generate S4 classes and functions for a specific table in the data base}
\usage{
\S4method{generateClass}{RangoSQLiteConnection}(dbc, tableName)
}
\arguments{
\item{dbc}{RangoPostgresConnection}

\item{tableName}{name of the table}
}
\value{
string containing the class definition
}
\description{
generate S4 classes and functions for a specific table in the data base
}
\author{
Willem Ligtenberg
}

