% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{generateClass}
\alias{generateClass}
\title{generate S4 classes and functions for a specific table in the data base}
\usage{
generateClass(dbc, tableName)
}
\arguments{
\item{dbc}{RangoConnection}

\item{tableName}{name of the table}
}
\value{
string containing the class definition
}
\description{
generate S4 classes and functions for a specific table in the data base
}
\author{
Willem Ligtenberg
}

