% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getConstraints,RangoObject-method}
\alias{getConstraints,RangoObject-method}
\title{Return part of a SQL statetement that specifies the constraints}
\usage{
\S4method{getConstraints}{RangoObject}(object, dbc, sql = TRUE,
  constraints = list())
}
\value{
String to insert in a SQL statement
}
\description{
Return part of a SQL statetement that specifies the constraints
}
\author{
Willem Ligtenberg
}

