% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getConstraints}
\alias{getConstraints}
\title{Return part of a SQL statetement that specifies the constraints}
\usage{
getConstraints(object, dbc, sql = TRUE, constraints = list())
}
\arguments{
\item{object}{object to inspect}

\item{dbc}{RangoConnection to use}

\item{sql}{Boolean, return a sql statement (TRUE), or list (FALSE)}

\item{clauses}{List of clauses already identified}
}
\value{
String to insert in a SQL statement
}
\description{
Return part of a SQL statetement that specifies the constraints
}
\author{
Willem Ligtenberg
}

