% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getJoinClauses,RangoObject-method}
\alias{getJoinClauses,RangoObject-method}
\title{Return part of a SQL statetement that specifies the join clauses}
\usage{
\S4method{getJoinClauses}{RangoObject}(object, dbc, sql = TRUE,
  clauses = list())
}
\value{
String to insert in a SQL statement
}
\description{
Return part of a SQL statetement that specifies the join clauses
}
\author{
Willem Ligtenberg
}

