% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rango.R
\docType{methods}
\name{getTableStatement,list-method}
\alias{getTableStatement,list-method}
\title{Return part of a SQL statetement that specifies the tables}
\usage{
\S4method{getTableStatement}{list}(object, dbc, sql = TRUE,
  tables = list())
}
\value{
String to insert in a SQL statement
}
\description{
Return part of a SQL statetement that specifies the tables
}
\author{
Willem Ligtenberg
}
