% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getTableStatement,RangoObject-method}
\alias{getTableStatement,RangoObject-method}
\title{Return part of a SQL statetement that specifies the tables}
\usage{
\S4method{getTableStatement}{RangoObject}(object, dbc, sql = TRUE,
  tables = list())
}
\value{
String to insert in a SQL statement
}
\description{
Return part of a SQL statetement that specifies the tables
}
\author{
Willem Ligtenberg
}

