% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getTableStatement}
\alias{getTableStatement}
\title{Return part of a SQL statetement that specifies the tables}
\usage{
getTableStatement(object, dbc, sql = TRUE, tables = list())
}
\arguments{
\item{object}{object to inspect}

\item{dbc}{RangoConnection to use}

\item{sql}{Boolean, return a sql statement (TRUE), or list (FALSE)}

\item{tables}{List of tables already identified}
}
\value{
String to insert in a SQL statement
}
\description{
Return part of a SQL statetement that specifies the tables
}
\author{
Willem Ligtenberg
}

