% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%>=\%}
\alias{\%>=\%}
\title{Greater than equals function to allow specification arguments where you want some
parameter to be greater than or equal to the following value.}
\usage{
var \%>=\% x
}
\arguments{
\item{var}{the name of the parameter}

\item{x}{the value of the parameter}
}
\value{
a string that performs the correct sql query
}
\description{
Greater than equals function to allow specification arguments where you want some
parameter to be greater than or equal to the following value.
}
\author{
Willem Ligtenberg
}

