% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rango.R
\name{\%<=>\%}
\alias{\%<=>\%}
\title{Between function to allow specification arguments where you want some 
parameter to be between the values specified in the vector.}
\usage{
var \%<=>\% x
}
\arguments{
\item{var}{the name of the parameter}

\item{x}{vector specifying the begin and end value}
}
\value{
a string that performs the correct sql query
}
\description{
Between function to allow specification arguments where you want some 
parameter to be between the values specified in the vector.
}
\author{
Willem Ligtenberg
}
