% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R, R/postgres.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{\S4method{initialize}{RangoSQLiteConnection}}
\alias{\S4method{initialize}{RangoPostgresConnection}}
\title{Initialize the RangoSQLiteConnection class}
\usage{
\S4method{initialize}{RangoSQLiteConnection}(.Object, dbname, cache)

\S4method{initialize}{RangoPostgresConnection}(.Object, host, dbname, port,
  user, password, cache)
}
\arguments{
\item{.Object}{RangoSQLiteConnection object}

\item{dbname}{database name}

\item{cache}{boolean which specifies if we use a local cache}

\item{host}{host name}

\item{port}{port on which the database server is listening}

\item{user}{user}

\item{password}{password}

\item{.Object}{RangoPostgresConnection object}

\item{dbname}{database name}

\item{cache}{boolean which specifies if we use a local cache}
}
\value{
RangoConnection object

RangoConnection object
}
\description{
Initialize the RangoSQLiteConnection class

Initialize the RangoPostgresConnection class
}
\author{
Willem Ligtenberg

Willem Ligtenberg
}
