% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\docType{methods}
\name{listTables,RangoSQLiteConnection-method}
\alias{listTables,RangoSQLiteConnection-method}
\title{list the tables for the current data base}
\usage{
\S4method{listTables}{RangoSQLiteConnection}(dbc)
}
\value{
vector containing table names in the data base
}
\description{
list the tables for the current data base
}
\author{
Willem Ligtenberg
}
