% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateClasses.R
\docType{methods}
\name{listTables}
\alias{listTables}
\title{list the tables for the current data base}
\usage{
listTables(dbc)
}
\arguments{
\item{dbc}{RangoConnection}
}
\value{
vector containing table names in the data base
}
\description{
list the tables for the current data base
}
\author{
Willem Ligtenberg
}
