% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateClasses.R
\name{loadClasses}
\alias{loadClasses}
\title{generate S4 classes and functions for a specific table in the data base and
loads them into memory}
\usage{
loadClasses(dbc)
}
\arguments{
\item{dbc}{RangoConnection}
}
\description{
generate S4 classes and functions for a specific table in the data base and
loads them into memory
}
\author{
Willem Ligtenberg
}
