% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rango.R
\docType{methods}
\name{nonEmptySlots,RangoObject-method}
\alias{nonEmptySlots,RangoObject-method}
\title{detect non empty slots of an object}
\usage{
\S4method{nonEmptySlots}{RangoObject}(object, dbc)
}
\value{
vector of non empty slots
}
\description{
detect non empty slots of an object
}
\author{
Willem Ligtenberg
}
