% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rango.R
\docType{methods}
\name{nonEmptySlots}
\alias{nonEmptySlots}
\title{detect non empty slots of an object}
\usage{
nonEmptySlots(object, dbc)
}
\arguments{
\item{object}{object to inspect}

\item{dbc}{RangoConnection to use}
}
\value{
vector of non empty slots
}
\description{
detect non empty slots of an object
}
\author{
Willem Ligtenberg
}
