% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateClasses.R
\name{optionalField}
\alias{optionalField}
\title{determine the optional fields of a table}
\usage{
optionalField(columnData)
}
\arguments{
\item{columnData}{data of a specific column of the data table}
}
\value{
string with the optional fields concatenated
}
\description{
determine the optional fields of a table
}
\author{
Willem Ligtenberg
}
