% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangoPackage.R
\docType{package}
\name{Rango}
\alias{Rango}
\alias{Rango-package}
\title{Rango: creates S4 objects from a database}
\description{
Willem Ligtenberg: willem.ligtenberg@openanalytics.eu
}
\details{
https://github.com/openanalytics/Rango
}
