% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangoConnection.R
\docType{class}
\name{RangoConnection-class}
\alias{RangoConnection-class}
\title{Object for the connection to the database}
\description{
Object for the connection to the database
}
\author{
Willem Ligtenberg
}
