% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangoConnection.R
\name{rangoConnection}
\alias{rangoConnection}
\title{Create connection to data base}
\usage{
rangoConnection(dbname, host = NA, port = 5432, user = NA,
  password = NA, type = "PostgreSQL", cache = FALSE,
  logLevel = loglevels["INFO"])
}
\arguments{
\item{dbname}{database name}

\item{host}{host name}

\item{port}{port on which the database server is listening}

\item{user}{user}

\item{password}{password}

\item{type}{database type (PostgreSQL)}

\item{cache}{boolean that specifies if caching should be used}
}
\value{
Connection object
}
\description{
Create connection to data base
}
\author{
Willem Ligtenberg
}
