% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\docType{class}
\name{RangoPostgresConnection-class}
\alias{RangoPostgresConnection-class}
\title{Object for the connection to the Postgres database}
\description{
Object for the connection to the Postgres database
}
\author{
Willem Ligtenberg
}
