% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\docType{class}
\name{RangoSQLiteConnection-class}
\alias{RangoSQLiteConnection-class}
\title{Object for the connection to the SQLite database}
\description{
Object for the connection to the SQLite database
}
\author{
Willem Ligtenberg
}
