% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\docType{methods}
\name{reconnect,RangoPostgresConnection-method}
\alias{reconnect,RangoPostgresConnection-method}
\title{Reconnect to the database without losing the local cache}
\usage{
\S4method{reconnect}{RangoPostgresConnection}(dbc)
}
\arguments{
\item{dbc}{Old RangoPostgresConnection object}
}
\value{
renewed RangoPostgresConnection object
}
\description{
Reconnect to the database without losing the local cache
}
\author{
Willem Ligtenberg
}
