% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\docType{methods}
\name{reconnect,RangoSQLiteConnection-method}
\alias{reconnect,RangoSQLiteConnection-method}
\title{Reconnect to the database without losing the local cache}
\usage{
\S4method{reconnect}{RangoSQLiteConnection}(dbc)
}
\arguments{
\item{dbc}{Old RangoSQLiteConnection object}
}
\value{
renewed RangoSQLiteConnection object
}
\description{
Reconnect to the database without losing the local cache
}
\author{
Willem Ligtenberg
}
