% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateClasses.R
\name{requiredFields}
\alias{requiredFields}
\title{determine the required fields of a table}
\usage{
requiredFields(columnData)
}
\arguments{
\item{columnData}{data of a specific column of the data table}
}
\value{
string with the required fields concatenated
}
\description{
determine the required fields of a table
}
\author{
Willem Ligtenberg
}
