% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rEsc}
\alias{rEsc}
\title{escape some quotes for R string}
\usage{
rEsc(s)
}
\arguments{
\item{s}{String to escape}
}
\value{
escaped string
}
\description{
escape some quotes for R string
}
\author{
Willem Ligtenberg
}
