% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{retrieve,list-method}
\alias{retrieve,list-method}
\title{query the data base for objects with the specific attributes in
the object that is passed.}
\usage{
\S4method{retrieve}{list}(object, dbc, create = FALSE,
  returnType = "RangoObject")
}
\value{
returns (a list of) object with all the
information from the data base (or a data.frame if chosen)
}
\description{
query the data base for objects with the specific attributes in
the object that is passed.
}
\author{
Willem Ligtenberg
}

