% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rango.R
\docType{methods}
\name{retrieve,RangoObject-method}
\alias{retrieve,RangoObject-method}
\title{query the data base for objects with the specific attributes in 
the object that is passed.}
\usage{
\S4method{retrieve}{RangoObject}(object, dbc, create = FALSE,
  returnType = "RangoObject")
}
\value{
returns (a list of) object with all the 
information from the data base (or a data.frame if chosen)
}
\description{
query the data base for objects with the specific attributes in 
the object that is passed.
}
\author{
Willem Ligtenberg
}
