% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{rollback,RangoPostgresConnection-method}
\alias{rollback,RangoPostgresConnection-method}
\title{Roll back the changes in the current transaction}
\usage{
\S4method{rollback}{RangoPostgresConnection}(object)
}
\description{
Roll back the changes in the current transaction
}
\author{
Willem Ligtenberg
}

