% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\docType{methods}
\name{rollback,RangoSQLiteConnection-method}
\alias{rollback,RangoSQLiteConnection-method}
\title{Roll back the changes in the current transaction}
\usage{
\S4method{rollback}{RangoSQLiteConnection}(object)
}
\description{
Roll back the changes in the current transaction
}
\author{
Willem Ligtenberg
}
