% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangoConnection.R
\name{rollback}
\alias{rollback}
\title{Roll back the changes in the current transaction}
\usage{
rollback(object)
}
\arguments{
\item{object}{the database connection to use}
}
\description{
Roll back the changes in the current transaction
}
\author{
Willem Ligtenberg
}
