% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqlEsc}
\alias{sqlEsc}
\title{escape some nasties in SQL strings}
\usage{
sqlEsc(s)
}
\arguments{
\item{s}{String to escape}
}
\value{
escaped string
}
\description{
escape some nasties in SQL strings
}
\author{
Willem Ligtenberg
}
