% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rango.R
\docType{methods}
\name{store,list,RangoPostgresConnection-method}
\alias{store,list,RangoPostgresConnection-method}
\title{store the list to the data base}
\usage{
\S4method{store}{list,RangoPostgresConnection}(object, dbc,
  force = FALSE, returnType = "RangoObject")
}
\value{
returns (a list of) object with all the
information from the data base (or a data.frame if chosen)
}
\description{
store the list to the data base
}
\author{
Willem Ligtenberg
}
