% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{store}
\alias{store}
\title{store the object to the data base}
\usage{
store(object, dbc, force = FALSE, returnType = "RangoObject")
}
\arguments{
\item{object}{object to store}

\item{dbc}{RangoConnection}

\item{force}{Should we create a new row even when
something like this already exists}

\item{returnType}{How should the result be returned? Default is as
RangoObject, None returns nothing (only useful for store), data.frame
returns a data.frame}
}
\value{
returns (a list of) object with all the
information from the data base (or a data.frame if chosen)
}
\description{
store the object to the data base
}
\author{
Willem Ligtenberg
}

