% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{stringRep}
\alias{stringRep}
\title{Generate a string representation for the object}
\usage{
stringRep(object, dbc)
}
\arguments{
\item{object}{object to represent as a string}

\item{dbc}{RangoConnection}
}
\value{
string representation of the object
}
\description{
Generate a string representation for the object
}
\author{
Willem Ligtenberg
}

