% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unlockEnvironment}
\alias{unlockEnvironment}
\title{unlock an R package on the R search() path}
\usage{
unlockEnvironment(env)
}
\arguments{
\item{env}{environment
e.g as.environment(2)
e.g getNamespace("Rango"))}
}
\description{
Note: R package base is unlocked by design
}
