% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{beginTransaction,RangoPostgresConnection-method}
\alias{beginTransaction,RangoPostgresConnection-method}
\title{Begin an new transaction}
\usage{
\S4method{beginTransaction}{RangoPostgresConnection}(object)
}
\arguments{
\item{object}{the database connection to use}
}
\description{
Begin an new transaction
}
\author{
Willem Ligtenberg
}

