% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beginTransaction}
\alias{beginTransaction}
\title{Begin an new transaction}
\usage{
beginTransaction(object)
}
\arguments{
\item{object}{the database connection to use}
}
\description{
Begin an new transaction
}
\author{
Willem Ligtenberg
}

