% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{commit,RangoSQLiteConnection-method}
\alias{commit,RangoSQLiteConnection-method}
\title{Commit the changes in the current transaction}
\usage{
\S4method{commit}{RangoSQLiteConnection}(object)
}
\arguments{
\item{object}{the database connection to use}
}
\description{
Commit the changes in the current transaction
}
\author{
Willem Ligtenberg
}

