% Generated by roxygen2 (4.0.2): do not edit by hand
\name{forceAssignMyNamespace}
\alias{forceAssignMyNamespace}
\title{Function to force the assignment of new items in the Rango namespace
Adapted from Rcpp}
\usage{
forceAssignMyNamespace(x, value)
}
\arguments{
\item{x}{}

\item{value}{}
}
\description{
Function to force the assignment of new items in the Rango namespace
Adapted from Rcpp
}
\author{
Willem Ligtenberg
}

