% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateClasses}
\alias{generateClasses}
\title{generate S4 classes and functions for a specific table in the data base}
\usage{
generateClasses(dbc = rangoConnection(), fileName = "dbClasses.R")
}
\arguments{
\item{dbc}{RangoConnection}

\item{fileName}{file name to write the classes to}
}
\description{
generate S4 classes and functions for a specific table in the data base
}
\author{
Willem Ligtenberg
}

