% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getPrimaryKey,RangoSQLiteConnection-method}
\alias{getPrimaryKey,RangoSQLiteConnection-method}
\title{Get the primary key for a table}
\usage{
\S4method{getPrimaryKey}{RangoSQLiteConnection}(dbc, tableName)
}
\arguments{
\item{dbc}{database connection to use}

\item{tableName}{name of the table}
}
\value{
string that contains the primary key of the table
}
\description{
Get the primary key for a table
}
\author{
Willem Ligtenberg
}

