% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%like\%}
\alias{\%like\%}
\title{Like function to allow specification arguments for partial matching.}
\usage{
var \%like\% x
}
\arguments{
\item{var}{the name of the parameter}

\item{x}{vector string to match}
}
\value{
a string that performs the correct sql query
}
\description{
Like function to allow specification arguments for partial matching.
}
\author{
Willem Ligtenberg
}

