% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize}
\alias{\S4method{initialize}{RangoPostgresConnection}}
\alias{\S4method{initialize}{RangoSQLiteConnection}}
\alias{initialize}
\title{Initialize the RangoPostgresConnection class}
\usage{
\S4method{initialize}{RangoPostgresConnection}(.Object, host, dbname, user,
  password, cache)

\S4method{initialize}{RangoSQLiteConnection}(.Object, dbname, cache)
}
\arguments{
\item{.Object}{RangoPostgresConnection object}

\item{host}{host name}

\item{dbname}{database name}

\item{user}{user}

\item{password}{password}

\item{cache}{boolean which specifies if we use a local cache}

\item{.Object}{RangoSQLiteConnection object}

\item{dbname}{database name}

\item{cache}{boolean which specifies if we use a local cache}
}
\value{
RangoConnection object

RangoConnection object
}
\description{
Initialize the RangoPostgresConnection class

Initialize the RangoSQLiteConnection class
}
\author{
Willem Ligtenberg

Willem Ligtenberg
}

