% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nonEmptySlots,RangoObject-method}
\alias{nonEmptySlots,RangoObject-method}
\title{detect non empty slots of an object}
\usage{
\S4method{nonEmptySlots}{RangoObject}(object, dbc)
}
\value{
vector of non empty slots
}
\description{
detect non empty slots of an object
}
\author{
Willem Ligtenberg
}

