% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optionalField}
\alias{optionalField}
\title{determine the optional fields of a table}
\usage{
optionalField(columnData)
}
\arguments{
\item{columnData}{data of a specific column of the data table}
}
\value{
string with the optional fields concatenated
}
\description{
determine the optional fields of a table
}
\author{
Willem Ligtenberg
}

