% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rangoConnection}
\alias{rangoConnection}
\title{Create connection to data base}
\usage{
rangoConnection(dbname, host = NA, user = NA, password = NA,
  type = "PostgreSQL", cache = FALSE, logLevel = loglevels["INFO"])
}
\arguments{
\item{dbname}{database name}

\item{host}{host name}

\item{user}{user}

\item{password}{password}

\item{type}{database type (PostgreSQL)}

\item{cache}{boolean that specifies if caching should be used}
}
\value{
Connection object
}
\description{
Create connection to data base
}
\author{
Willem Ligtenberg
}

