% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{RangoPostgresConnection-class}
\alias{RangoPostgresConnection-class}
\title{Object for the connection to the Postgres database}
\description{
Object for the connection to the Postgres database
}
\author{
Willem Ligtenberg
}

