% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reconnect}
\alias{reconnect}
\title{Reconnect to the database without losing the local cache}
\usage{
reconnect(dbc)
}
\arguments{
\item{dbc}{Old RangoConnection object}
}
\description{
Reconnect to the database without losing the local cache
}
\author{
Willem Ligtenberg
}

