% Generated by roxygen2 (4.0.2): do not edit by hand
\name{requiredFields}
\alias{requiredFields}
\title{determine the required fields of a table}
\usage{
requiredFields(columnData)
}
\arguments{
\item{columnData}{data of a specific column of the data table}
}
\value{
string with the required fields concatenated
}
\description{
determine the required fields of a table
}
\author{
Willem Ligtenberg
}

