% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{retrieve}
\alias{retrieve}
\title{query the data base for objects with the specific attributes in
the object that is passed.}
\usage{
retrieve(object, dbc, create = TRUE, returnType = "RangoObject")
}
\arguments{
\item{object}{object to retrieve}

\item{dbc}{RangoConnection}

\item{create}{Should a new entry be made in the data base if none exist}

\item{returnType}{How should the result be returned? Default is as
RangoObject, None returns nothing (only useful for store), data.frame
returns a data.frame}
}
\value{
returns (a list of) object with all the
information from the data base (or a data.frame if chosen)
}
\description{
query the data base for objects with the specific attributes in
the object that is passed.
}
\author{
Willem Ligtenberg
}

