% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{store,RangoObject,RangoPostgresConnection-method}
\alias{store,RangoObject,RangoPostgresConnection-method}
\title{store the object to the data base}
\usage{
\S4method{store}{RangoObject,RangoPostgresConnection}(object, dbc,
  force = FALSE, returnType = "RangoObject")
}
\value{
returns (a list of) object with all the
information from the data base (or a data.frame if chosen)
}
\description{
store the object to the data base
}
\author{
Willem Ligtenberg
}

