% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{stringRep,RangoObject-method}
\alias{stringRep,RangoObject-method}
\title{Generate a string representation for the object}
\usage{
\S4method{stringRep}{RangoObject}(object, dbc)
}
\value{
string representation of the object
}
\description{
Generate a string representation for the object
}
\author{
Willem Ligtenberg
}

